+++ snapshots/alias.go
@@ -2,7 +2,15 @@ package snapshots
 
 import "github.com/cosmos/cosmos-sdk/snapshots/types"
 
+const (
+	CurrentFormat = types.CurrentFormat
+)
+
+var (
+	ErrUnknownFormat = types.ErrUnknownFormat
+)
+
 type (
-	Snapshot = types.Snapshot
-	Chunk    = types.Chunk
+	Snapshotter = types.Snapshotter
+	Snapshot    = types.Snapshot
 )