+++ snapshots/types/error.go
@@ -1,0 +1,8 @@
+package types
+
+import "errors"
+
+var (
+	// ErrUnknownFormat is returned when an unknown format is used.
+	ErrUnknownFormat = errors.New("unknown snapshot format")
+)