+++ snapshots/types/types.pb.go
@@ -24,9 +24,9 @@ const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
 // Snapshot contains snapshot metadata.
 type Snapshot struct {
-	Height uint64   `protobuf:"varint,1,opt,name=height,proto3" json:"height,omitempty"`
-	Format uint32   `protobuf:"varint,2,opt,name=format,proto3" json:"format,omitempty"`
-	Chunks []*Chunk `protobuf:"bytes,3,rep,name=chunks,proto3" json:"chunks,omitempty"`
+	Height      uint64   `protobuf:"varint,1,opt,name=height,proto3" json:"height,omitempty"`
+	Format      uint32   `protobuf:"varint,2,opt,name=format,proto3" json:"format,omitempty"`
+	ChunkHashes [][]byte `protobuf:"bytes,3,rep,name=chunk_hashes,json=chunkHashes,proto3" json:"chunk_hashes,omitempty"`
 }
 
 func (m *Snapshot) Reset()         { *m = Snapshot{} }
@@ -76,81 +76,34 @@ func (m *Snapshot) GetFormat() uint32 {
 	return 0
 }
 
-func (m *Snapshot) GetChunks() []*Chunk {
+func (m *Snapshot) GetChunkHashes() [][]byte {
 	if m != nil {
-		return m.Chunks
-	}
-	return nil
-}
-
-// Chunk contains snapshot chunk metadata.
-type Chunk struct {
-	Hash []byte `protobuf:"bytes,1,opt,name=hash,proto3" json:"hash,omitempty"`
-}
-
-func (m *Chunk) Reset()         { *m = Chunk{} }
-func (m *Chunk) String() string { return proto.CompactTextString(m) }
-func (*Chunk) ProtoMessage()    {}
-func (*Chunk) Descriptor() ([]byte, []int) {
-	return fileDescriptor_03ed7742cffb8ebd, []int{1}
-}
-func (m *Chunk) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *Chunk) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_Chunk.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalToSizedBuffer(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *Chunk) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Chunk.Merge(m, src)
-}
-func (m *Chunk) XXX_Size() int {
-	return m.Size()
-}
-func (m *Chunk) XXX_DiscardUnknown() {
-	xxx_messageInfo_Chunk.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Chunk proto.InternalMessageInfo
-
-func (m *Chunk) GetHash() []byte {
-	if m != nil {
-		return m.Hash
+		return m.ChunkHashes
 	}
 	return nil
 }
 
 func init() {
 	proto.RegisterType((*Snapshot)(nil), "cosmos_sdk.snapshots.v1.Snapshot")
-	proto.RegisterType((*Chunk)(nil), "cosmos_sdk.snapshots.v1.Chunk")
 }
 
 func init() { proto.RegisterFile("snapshots/types/types.proto", fileDescriptor_03ed7742cffb8ebd) }
 
 var fileDescriptor_03ed7742cffb8ebd = []byte{
-	// 213 bytes of a gzipped FileDescriptorProto
+	// 194 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0x92, 0x2e, 0xce, 0x4b, 0x2c,
 	0x28, 0xce, 0xc8, 0x2f, 0x29, 0xd6, 0x2f, 0xa9, 0x2c, 0x48, 0x85, 0x92, 0x7a, 0x05, 0x45, 0xf9,
 	0x25, 0xf9, 0x42, 0xe2, 0xc9, 0xf9, 0xc5, 0xb9, 0xf9, 0xc5, 0xf1, 0xc5, 0x29, 0xd9, 0x7a, 0x70,
-	0x75, 0x7a, 0x65, 0x86, 0x4a, 0x45, 0x5c, 0x1c, 0xc1, 0x50, 0xbe, 0x90, 0x18, 0x17, 0x5b, 0x46,
+	0x75, 0x7a, 0x65, 0x86, 0x4a, 0xb1, 0x5c, 0x1c, 0xc1, 0x50, 0xbe, 0x90, 0x18, 0x17, 0x5b, 0x46,
 	0x6a, 0x66, 0x7a, 0x46, 0x89, 0x04, 0xa3, 0x02, 0xa3, 0x06, 0x4b, 0x10, 0x94, 0x07, 0x12, 0x4f,
-	0xcb, 0x2f, 0xca, 0x4d, 0x2c, 0x91, 0x60, 0x52, 0x60, 0xd4, 0xe0, 0x0d, 0x82, 0xf2, 0x84, 0xcc,
-	0xb8, 0xd8, 0x92, 0x33, 0x4a, 0xf3, 0xb2, 0x8b, 0x25, 0x98, 0x15, 0x98, 0x35, 0xb8, 0x8d, 0xe4,
-	0xf4, 0x70, 0xd8, 0xa2, 0xe7, 0x0c, 0x52, 0x16, 0x04, 0x55, 0xad, 0x24, 0xcd, 0xc5, 0x0a, 0x16,
-	0x10, 0x12, 0xe2, 0x62, 0xc9, 0x48, 0x2c, 0xce, 0x00, 0x5b, 0xc7, 0x13, 0x04, 0x66, 0x3b, 0xb9,
-	0x9d, 0x78, 0x24, 0xc7, 0x78, 0xe1, 0x91, 0x1c, 0xe3, 0x83, 0x47, 0x72, 0x8c, 0x13, 0x1e, 0xcb,
-	0x31, 0x5c, 0x78, 0x2c, 0xc7, 0x70, 0xe3, 0xb1, 0x1c, 0x43, 0x94, 0x4e, 0x7a, 0x66, 0x49, 0x46,
-	0x69, 0x92, 0x5e, 0x72, 0x7e, 0xae, 0x3e, 0xc4, 0x22, 0x28, 0xa5, 0x5b, 0x9c, 0x92, 0xad, 0x8f,
-	0xe6, 0xfb, 0x24, 0x36, 0xb0, 0xc7, 0x8d, 0x01, 0x01, 0x00, 0x00, 0xff, 0xff, 0x27, 0x94, 0xf7,
-	0x05, 0x17, 0x01, 0x00, 0x00,
+	0xcb, 0x2f, 0xca, 0x4d, 0x2c, 0x91, 0x60, 0x52, 0x60, 0xd4, 0xe0, 0x0d, 0x82, 0xf2, 0x84, 0x14,
+	0xb9, 0x78, 0x92, 0x33, 0x4a, 0xf3, 0xb2, 0xe3, 0x33, 0x12, 0x8b, 0x33, 0x52, 0x8b, 0x25, 0x98,
+	0x15, 0x98, 0x35, 0x78, 0x82, 0xb8, 0xc1, 0x62, 0x1e, 0x60, 0x21, 0x27, 0xb7, 0x13, 0x8f, 0xe4,
+	0x18, 0x2f, 0x3c, 0x92, 0x63, 0x7c, 0xf0, 0x48, 0x8e, 0x71, 0xc2, 0x63, 0x39, 0x86, 0x0b, 0x8f,
+	0xe5, 0x18, 0x6e, 0x3c, 0x96, 0x63, 0x88, 0xd2, 0x49, 0xcf, 0x2c, 0xc9, 0x28, 0x4d, 0xd2, 0x4b,
+	0xce, 0xcf, 0xd5, 0x87, 0x38, 0x0e, 0x4a, 0xe9, 0x16, 0xa7, 0x64, 0xeb, 0xa3, 0xf9, 0x25, 0x89,
+	0x0d, 0xec, 0x0d, 0x63, 0x40, 0x00, 0x00, 0x00, 0xff, 0xff, 0x7f, 0x74, 0x08, 0x41, 0xe5, 0x00,
+	0x00, 0x00,
 }
 
 func (m *Snapshot) Marshal() (dAtA []byte, err error) {
@@ -173,16 +126,11 @@ func (m *Snapshot) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Chunks) > 0 {
-		for iNdEx := len(m.Chunks) - 1; iNdEx >= 0; iNdEx-- {
-			{
-				size, err := m.Chunks[iNdEx].MarshalToSizedBuffer(dAtA[:i])
-				if err != nil {
-					return 0, err
-				}
-				i -= size
-				i = encodeVarintTypes(dAtA, i, uint64(size))
-			}
+	if len(m.ChunkHashes) > 0 {
+		for iNdEx := len(m.ChunkHashes) - 1; iNdEx >= 0; iNdEx-- {
+			i -= len(m.ChunkHashes[iNdEx])
+			copy(dAtA[i:], m.ChunkHashes[iNdEx])
+			i = encodeVarintTypes(dAtA, i, uint64(len(m.ChunkHashes[iNdEx])))
 			i--
 			dAtA[i] = 0x1a
 		}
@@ -196,36 +144,6 @@ func (m *Snapshot) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 		i = encodeVarintTypes(dAtA, i, uint64(m.Height))
 		i--
 		dAtA[i] = 0x8
-	}
-	return len(dAtA) - i, nil
-}
-
-func (m *Chunk) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalToSizedBuffer(dAtA[:size])
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *Chunk) MarshalTo(dAtA []byte) (int, error) {
-	size := m.Size()
-	return m.MarshalToSizedBuffer(dAtA[:size])
-}
-
-func (m *Chunk) MarshalToSizedBuffer(dAtA []byte) (int, error) {
-	i := len(dAtA)
-	_ = i
-	var l int
-	_ = l
-	if len(m.Hash) > 0 {
-		i -= len(m.Hash)
-		copy(dAtA[i:], m.Hash)
-		i = encodeVarintTypes(dAtA, i, uint64(len(m.Hash)))
-		i--
-		dAtA[i] = 0xa
 	}
 	return len(dAtA) - i, nil
 }
@@ -253,24 +171,11 @@ func (m *Snapshot) Size() (n int) {
 	if m.Format != 0 {
 		n += 1 + sovTypes(uint64(m.Format))
 	}
-	if len(m.Chunks) > 0 {
-		for _, e := range m.Chunks {
-			l = e.Size()
+	if len(m.ChunkHashes) > 0 {
+		for _, b := range m.ChunkHashes {
+			l = len(b)
 			n += 1 + l + sovTypes(uint64(l))
 		}
-	}
-	return n
-}
-
-func (m *Chunk) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	l = len(m.Hash)
-	if l > 0 {
-		n += 1 + l + sovTypes(uint64(l))
 	}
 	return n
 }
@@ -350,94 +255,7 @@ func (m *Snapshot) Unmarshal(dAtA []byte) error {
 			}
 		case 3:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Chunks", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowTypes
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthTypes
-			}
-			postIndex := iNdEx + msglen
-			if postIndex < 0 {
-				return ErrInvalidLengthTypes
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.Chunks = append(m.Chunks, &Chunk{})
-			if err := m.Chunks[len(m.Chunks)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
-			iNdEx = postIndex
-		default:
-			iNdEx = preIndex
-			skippy, err := skipTypes(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthTypes
-			}
-			if (iNdEx + skippy) < 0 {
-				return ErrInvalidLengthTypes
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *Chunk) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowTypes
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= uint64(b&0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: Chunk: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: Chunk: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Hash", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field ChunkHashes", wireType)
 			}
 			var byteLen int
 			for shift := uint(0); ; shift += 7 {
@@ -464,10 +282,8 @@ func (m *Chunk) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Hash = append(m.Hash[:0], dAtA[iNdEx:postIndex]...)
-			if m.Hash == nil {
-				m.Hash = []byte{}
-			}
+			m.ChunkHashes = append(m.ChunkHashes, make([]byte, postIndex-iNdEx))
+			copy(m.ChunkHashes[len(m.ChunkHashes)-1], dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex